feat(emulate): export enum variants
