+++ Cargo.toml
@@ -21,7 +21,6 @@ default = ["emulation"]
 
 emulation = ["dep:typed-builder", "dep:brotli", "dep:flate2", "dep:zstd"]
 emulation-serde = ["dep:serde"]
-emulation-rand = ["emulation", "dep:strum_macros", "dep:strum"]
 emulation-compression = []
 
 tower-delay = ["dep:tower", "dep:pin-project-lite", "tokio/time"]
@@ -30,8 +29,6 @@ tower-delay = ["dep:tower", "dep:pin-project-lite", "tokio/time"]
 wreq = { git = "https://github.com/0x676e67/wreq.git", default-features = false }
 serde = { version = "1.0.228", features = ["derive"], optional = true }
 typed-builder = { version = "0.23.2", optional = true }
-strum = { version = "0.28.0", optional = true }
-strum_macros = { version = "0.28.0", optional = true }
 tokio = { version = "1.50.0", default-features = false, optional = true }
 tower = { version = "0.5.2", default-features = false, optional = true }
 pin-project-lite = { version = "0.2.17", optional = true }
